requirements.txt
Flask
Gunicorn
python-dotenv
langchain
pypdf==4.0.1
InstructorEmbedding==1.0.1
torch==2.2.1
tqdm==4.66.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pymongo
